Initial Blockout
Completed an initial blockout of each floor of the house using Godot's 3D gridmap system. No ceilings for each floor currently, as the stairs go too high to allow the floor above to double up as a ceiling. Will add a player controller and then return to this to find appropriate ceiling height once I've checked the scale.