Implemented requested changes
Only change not implemented from request is sound when unable to stand up.

Also, camera height is not actually that low - it is that far above the centre of the player. It should not need bringing up further, although it has been bumped up slightly in this while tweaking the camera.