Flashlight creation
Created a flashlight that is connected the players camera, it has a charge that goes down while active and begins to fade out at a certain point.

the flashlight can be toggled on and off with the right mouse button and there is a function that can be called with an input value that sets the charge of the flashlight.

Also replaced the lighting from the default to very dark. If you need to see increase the ambient light in the World Environment object.